Added Chart and fixed input data

+++ app/src/main/java/com/example/reservoir_near_you/model/MagasinX.kt
@@ -2,9 +2,9 @@ package com.example.reservoir_near_you.model
 
 data class MagasinX(
     val dato_Id: String,
-    val endring_fyllingsgrad: Double,
+    val endring_fyllingsgrad: Float,
     val fylling_TWh: Double,
-    val fyllingsgrad: Double,
+    val fyllingsgrad: Float,
     val fyllingsgrad_forrige_uke: Double,
     val id: Int,
     val iso_uke: Int,
